Add e2e tests for TP2 and TP3 workflows
- 14 TP2 tests: manual/auto calibration, export, simulation validation
- 12 TP3 tests: calibration for projection, Baskatong/Au Saumon projections
- Enhance page objects with export, auto-complete, and indicator methods
- Add catchment-specific fixtures (Leaf, Baskatong, Au Saumon)
- Fix int64 dtype bug in simulation API hydro params parsing
- Register pytest-timeout marker

+++ src/holmes/api/simulation.py
@@ -285,7 +285,7 @@ def _run_simulation(
     warmup_steps: int,
 ) -> tuple[npt.NDArray[np.float64], dict[str, float]]:
     hydro_simulate = hydro.get_model(cast(hydro.HydroModel, hydro_model))
-    hydro_params_ = np.array(list(hydro_params.values()))
+    hydro_params_ = np.array(list(hydro_params.values()), dtype=np.float64)
 
     if snow_model != "none":
         # These values are guaranteed to be non-None when snow_model is set
